+++ core-api/src/test/java/com/optimizely/ab/OptimizelyTest.java
@@ -40,7 +40,7 @@ import com.optimizely.ab.event.internal.payload.EventBatch;
 import com.optimizely.ab.internal.LogbackVerifier;
 import com.optimizely.ab.internal.ControlAttribute;
 import com.optimizely.ab.notification.*;
-import com.optimizely.ab.notification.decisionInfo.DecisionNotification;
+import com.optimizely.ab.notification.DecisionNotification;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.junit.Rule;
 import org.junit.Test;
@@ -66,14 +66,14 @@ import java.util.Map;
 import static com.optimizely.ab.config.ProjectConfigTestUtils.*;
 import static com.optimizely.ab.config.ValidProjectConfigV4.*;
 import static com.optimizely.ab.event.LogEvent.RequestMethod;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.FEATURE_ENABLED;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.FEATURE_KEY;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.SOURCE;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.SOURCE_EXPERIMENT_KEY;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.SOURCE_VARIATION_KEY;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.VARIABLE_KEY;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.VARIABLE_VALUE;
-import static com.optimizely.ab.notification.decisionInfo.FeatureVariableNotification.VARIABLE_TYPE;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.FEATURE_ENABLED;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.FEATURE_KEY;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.SOURCE;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.SOURCE_EXPERIMENT_KEY;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.SOURCE_VARIATION_KEY;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.VARIABLE_KEY;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.VARIABLE_VALUE;
+import static com.optimizely.ab.notification.DecisionNotification.FeatureVariableDecisionNotificationBuilder.VARIABLE_TYPE;
 import static java.util.Arrays.asList;
 import static junit.framework.TestCase.assertTrue;
 import static org.hamcrest.CoreMatchers.is;
